DSP-26 Where possible, push down filter predicates to Cassandra
Certain filter predicates can be translated to secondary index searches.
Where this is possible, identify the predicate and the indexed columns
to which they can be applied and pass them via the JobConf to the MR
tasks accessing Cassandra.

Instead of relying on Hive's type inference, use the validation class
defined in CF metadata to serialize the value of an index expression.
We need this as Hive & Cassandra types don't map 1:1, a Hive binary
may be a UUID, Varint, Decimal or Bytes type column in C*.
Added a set of UDFs to handle these mappings properly., plus 2 more
in the test source to sanity check composite column handling.
These are cribbed from examples:

http://www.datastax.com/support-forums/topic/how-do-you-use-composite-columns-from-cassandra-in-a-hive-table#post-346
https://github.com/edwardcapriolo/hive_cassandra_udfs/blob/master/src/main/java/com/jointhegrid/hivecasudfs/UDFCompositeToByteArray.java